Added image, border and basic position to Card
